Fix test database issue

+++ database/migrations/2016_09_12_211736_create_pages_table.php
@@ -57,8 +57,10 @@ class CreatePagesTable extends Migration
             }
         });
 
-        Schema::drop('skytz_blocks');
-        Schema::drop('skytz_strokes');
+        if (ImportTable::check()) {
+            Schema::drop('skytz_blocks');
+            Schema::drop('skytz_strokes');
+        }
     }
 
     /**
